Move utils to area of usage
