+++ src/index.ts
@@ -15,8 +15,9 @@ export * from "./types";
 
 export * from "./utils/filter";
 export * from "./utils/CommandError";
-export type { StreamMetadata } from "./utils/streamMetadata";
-export * from "./utils/systemStreams";
+
+export type { StreamMetadata } from "./streams/utils/streamMetadata";
+export * from "./streams/utils/systemStreams";
 export {
   PersistentSubscriptionSettings,
   PersistentSubscriptionToStreamSettings,
