+++ src/projections/getProjectionStatus.ts
@@ -5,11 +5,8 @@ import { StatisticsReq, StatisticsResp } from "../../generated/projections_pb";
 
 import { Client } from "../Client";
 import type { BaseOptions, ProjectionDetails } from "../types";
-import {
-  debug,
-  convertToCommandError,
-  convertGrpcProjectionDetails,
-} from "../utils";
+import { debug, convertToCommandError } from "../utils";
+import { mapGrpcProjectionDetails } from "./utils/mapGrpcProjectionDetails";
 
 export interface GetProjectionStatusOptions extends BaseOptions {}
 
@@ -74,7 +71,7 @@ Client.prototype.getProjectionStatus = async function (
 
         stream.on("data", (resp: StatisticsResp) => {
           if (!resp.hasDetails()) return;
-          projectionDetail = convertGrpcProjectionDetails(resp.getDetails()!);
+          projectionDetail = mapGrpcProjectionDetails(resp.getDetails()!);
         });
 
         stream.on("end", () => {
