+++ src/streams/utils/ReadStream.ts
@@ -3,15 +3,14 @@ import { Transform, TransformCallback, TransformOptions } from "stream";
 import type { ClientReadableStream, ServiceError } from "@grpc/grpc-js";
 import { Status } from "@grpc/grpc-js/build/src/constants";
 
-import type { ReadResp } from "../../generated/streams_pb";
+import type { ReadResp } from "../../../generated/streams_pb";
 
-import type { StreamingRead } from "../types";
-
+import type { StreamingRead } from "../../types";
 import {
   ConvertGrpcEvent,
   convertToCommandError,
   StreamNotFoundError,
-} from ".";
+} from "../../utils";
 
 type CreateGRPCStream = () => Promise<ClientReadableStream<ReadResp>>;
