+++ src/utils/index.ts
@@ -1,11 +1,7 @@
 export * from "./backpressuredWrite";
 export * from "./CommandError";
 export * from "./convertGrpcEvent";
-export * from "./convertGrpcProjectionDetails";
 export * from "./debug";
 export * from "./filter";
-export * from "./OneWaySubscription";
-export * from "./ReadStream";
-export * from "./systemStreams";
 export * from "./grpcStreamIdentifier";
 export * from "./grpcUUID";
